requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pynrrd>=0.4.3
transformers>=4.25.0
tqdm>=4.62.0
SimpleITK>=2.1.0
openpyxl>=3.0.9 